--- packages/hardhat/test/YouContract.ts
@@ -1,28 +1,0 @@
-import { expect } from "chai";
-import { ethers } from "hardhat";
-import { YourContract } from "../typechain-types";
-
-describe("YourContract", function () {
-  // We define a fixture to reuse the same setup in every test.
-
-  let yourContract: YourContract;
-  before(async () => {
-    const [owner] = await ethers.getSigners();
-    const yourContractFactory = await ethers.getContractFactory("YourContract");
-    yourContract = (await yourContractFactory.deploy(owner.address)) as YourContract;
-    await yourContract.deployed();
-  });
-
-  describe("Deployment", function () {
-    it("Should have the right message on deploy", async function () {
-      expect(await yourContract.greeting()).to.equal("Building Unstoppable Apps!!!");
-    });
-
-    it("Should allow setting a new message", async function () {
-      const newGreeting = "Learn Scaffold-ETH 2! :)";
-
-      await yourContract.setGreeting(newGreeting);
-      expect(await yourContract.greeting()).to.equal(newGreeting);
-    });
-  });
-});
