+++ packages/nextjs/components/Header.tsx
@@ -2,13 +2,7 @@ import React, { useCallback, useRef, useState } from "react";
 import Image from "next/image";
 import Link from "next/link";
 import { useRouter } from "next/router";
-import {
-  Bars3Icon,
-  BugAntIcon,
-  MagnifyingGlassIcon,
-  RocketLaunchIcon,
-  SparklesIcon,
-} from "@heroicons/react/24/outline";
+import { Bars3Icon, BugAntIcon, MagnifyingGlassIcon, RocketLaunchIcon } from "@heroicons/react/24/outline";
 import { FaucetButton, RainbowKitCustomConnectButton } from "~~/components/scaffold-eth";
 import { useOutsideClick } from "~~/hooks/scaffold-eth";
 
@@ -43,7 +37,10 @@ export const Header = () => {
   const navLinks = (
     <>
       <li>
-        <NavLink href="/">Home</NavLink>
+        <NavLink href="/">
+          <RocketLaunchIcon className="h-4 w-4" />
+          ShipStream
+        </NavLink>
       </li>
       <li>
         <NavLink href="/debug">
@@ -52,21 +49,9 @@ export const Header = () => {
         </NavLink>
       </li>
       <li>
-        <NavLink href="/example-ui">
-          <SparklesIcon className="h-4 w-4" />
-          Example UI
-        </NavLink>
-      </li>
-      <li>
         <NavLink href="/blockexplorer">
           <MagnifyingGlassIcon className="h-4 w-4" />
           Block Explorer
-        </NavLink>
-      </li>
-      <li>
-        <NavLink href="/shipstream">
-          <RocketLaunchIcon className="h-4 w-4" />
-          ShipStream
         </NavLink>
       </li>
     </>
