--- packages/nextjs/components/example-ui/assets/CopyIcon.tsx
@@ -1,24 +1,0 @@
-export const CopyIcon = ({ className }: { className: string }) => {
-  return (
-    <svg
-      className={className}
-      width="125"
-      height="117"
-      viewBox="0 0 125 117"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path d="M124.054 21.9365H116.774V145.99H124.054V21.9365Z" fill="#B5B5B5" />
-      <path d="M116.774 14.6089H109.446V21.9372H116.774V14.6089Z" fill="#B5B5B5" />
-      <path d="M109.446 7.3291H102.165V14.6095H109.446V7.3291Z" fill="#B5B5B5" />
-      <path
-        d="M109.436 36.4979H102.155V29.2415H94.8749V21.9372H87.5466V14.6089H0V153.295H109.436V36.4979ZM102.155 145.991H7.28039V21.9372H72.9475V51.1066H102.165L102.155 145.991Z"
-        fill="#B5B5B5"
-      />
-      <path d="M102.165 0H7.28052V7.32828H102.165V0Z" fill="#B5B5B5" />
-      <path d="M87.5566 94.8843H21.8894V102.213H87.5566V94.8843Z" fill="#B5B5B5" />
-      <path d="M87.5566 65.7144H21.8894V72.9947H87.5566V65.7144Z" fill="#B5B5B5" />
-      <path d="M51.0588 36.4971H21.8894V43.8254H51.0588V36.4971Z" fill="#B5B5B5" />
-    </svg>
-  );
-};
