--- packages/nextjs/pages/example-ui.tsx
@@ -1,25 +1,0 @@
-import type { NextPage } from "next";
-import { MetaHeader } from "~~/components/MetaHeader";
-import { ContractData } from "~~/components/example-ui/ContractData";
-import { ContractInteraction } from "~~/components/example-ui/ContractInteraction";
-
-const ExampleUI: NextPage = () => {
-  return (
-    <>
-      <MetaHeader
-        title="Example UI | Scaffold-ETH 2"
-        description="Example UI created with 🏗 Scaffold-ETH 2, showcasing some of its features."
-      >
-        {/* We are importing the font this way to lighten the size of SE2. */}
-        <link rel="preconnect" href="https://fonts.googleapis.com" />
-        <link href="https://fonts.googleapis.com/css2?family=Bai+Jamjuree&display=swap" rel="stylesheet" />
-      </MetaHeader>
-      <div className="grid lg:grid-cols-2 flex-grow" data-theme="exampleUi">
-        <ContractInteraction />
-        <ContractData />
-      </div>
-    </>
-  );
-};
-
-export default ExampleUI;
